+++ src/store/modules/connection.js
@@ -1,7 +1,7 @@
 export const connection = {
   namespaced: true,
   state: {
-    activeInstanceIdx: 0,
+    activeInstanceIdx: null,
     instances: [],
     elasticsearchAdapter: null
   },
@@ -18,7 +18,9 @@ export const connection = {
     },
     removeInstance (state, index) {
       state.instances.splice(index, 1)
-      if (index === state.activeInstanceIdx) {
+      if (state.instances.length === 0) {
+        state.activeInstanceIdx = null
+      } else if (index === state.activeInstanceIdx) {
         state.activeInstanceIdx = 0
       } else if (index < state.activeInstanceIdx) {
         state.activeInstanceIdx = state.activeInstanceIdx - 1